# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RTK Cloud Logger API
  version: 0.1.0
  description: >
    HTTP API for the RTK Cloud central service logging backend. Application
    services continue to emit logs to stdout/stderr; forwarders authenticate to
    this API and deliver normalized service log events.
servers:
  - url: http://127.0.0.1:18090
    description: Local development logger backend
security:
  - bearerAuth: []
paths:
  /v1/billing-usage/events:
    get:
      operationId: listBillingUsageInbox
      summary: Read a stable committed billing inbox horizon
      description: >
        Requires the dedicated billing credential. Records are ordered by
        inbox commit sequence, not producer time. An omitted cursor starts at
        the first retained record. Continue next_cursor until has_more is false;
        polling that final cursor opens a new committed horizon. Persist the
        cursor with validated facts. A complete horizon does not prove producer
        drainage or settlement. Store replacement or sequence gaps fail closed.
      parameters:
        - in: query
          name: cursor
          schema: { type: string, maxLength: 512 }
          description: Opaque store-bound cursor returned by the preceding page.
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        '200':
          description: Complete bounded page of committed receipts.
          content:
            application/json:
              schema:
                type: object
                required: [store_id, high_water, records, next_cursor, has_more]
                properties:
                  store_id: { type: string, pattern: '^[0-9a-f]{32}$' }
                  high_water: { type: string, pattern: '^[0-9]+$' }
                  next_cursor: { type: string }
                  has_more: { type: boolean }
                  records:
                    type: array
                    maxItems: 1000
                    items:
                      type: object
                      required: [sequence, content_sha256, event]
                      properties:
                        sequence: { type: string, pattern: '^[1-9][0-9]*$' }
                        content_sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
                        event: { $ref: '#/components/schemas/LogEvent' }
        '400': { description: Invalid limit or unsupported filter. }
        '401': { description: Dedicated billing credential required. }
        '405': { description: GET required. }
        '409': { description: Malformed cursor, wrong store or restored data behind cursor. }
        '503': { description: Inbox missing, closed, damaged or unavailable. }
  /healthz:
    get:
      summary: Health check
      description: Reports whether the logger backend HTTP process is alive.
      operationId: getHealth
      security: []
      responses:
        "204":
          description: Backend process is healthy.
  /metrics/prometheus:
    get:
      summary: Prometheus metrics
      description: Exposes low-cardinality Cloud Logger backend metrics for private Prometheus scraping.
      operationId: getPrometheusMetrics
      security: []
      responses:
        "200":
          description: Prometheus text exposition metrics.
          content:
            text/plain:
              schema:
                type: string
                example: |
                  # HELP rtk_cloud_logger_up Whether the Cloud Logger backend is serving metrics.
                  # TYPE rtk_cloud_logger_up gauge
                  rtk_cloud_logger_up 1
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/logs/ingest:
    post:
      summary: Ingest service log events
      description: >
        Accepts a batch of normalized service log events from a forwarder.
        Events are validated, redacted before persistence, and inserted
        idempotently by `event_id`. Deployments should enforce bounded request
        body size and maximum events per batch to protect the backend from
        accidental oversized forwarder uploads. The reference handler defaults
        to a 10 MiB request body limit and 1000 events per batch.
      operationId: ingestLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestRequest"
            examples:
              singleEvent:
                summary: Single journald event
                value:
                  events:
                    - event_id: 6f8d9cf1b41e4f4aa1e76f0a7c6328e8b8dbd48c2d5724ecacda54f1e0f0d942
                      ts: "2026-06-01T01:02:03Z"
                      level: info
                      msg: http request
                      service: video-cloud-api
                      env: staging
                      version: v0.1.0
                      host: api-1
                      unit: video-cloud-api.service
                      source: journald
                      trace_id: trace-1
                      request_id: request-1
                      operation_id: operation-1
                      device_id: device-1
                      org_id: org-1
                      user_id: user-1
                      component: http
                      fields:
                        method: GET
                        path: /v1/health
                        status: 200
                        duration_ms: 3.4
      responses:
        "202":
          description: Batch was processed; inspect per-event result statuses.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResponse"
              examples:
                acceptedAndDuplicate:
                  value:
                    results:
                      - event_id: 6f8d9cf1b41e4f4aa1e76f0a7c6328e8b8dbd48c2d5724ecacda54f1e0f0d942
                        status: accepted
                      - event_id: 6f8d9cf1b41e4f4aa1e76f0a7c6328e8b8dbd48c2d5724ecacda54f1e0f0d942
                        status: duplicate
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
  /v1/logs:
    get:
      summary: Query service log events
      description: >
        Returns stored service log events matching the provided filters.
        High-cardinality identifiers remain query fields and are not default
        Loki-style labels. Results are bounded by `limit` and ordered by event
        timestamp using `order`.
      operationId: queryLogs
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
        - $ref: "#/components/parameters/EventID"
        - $ref: "#/components/parameters/Env"
        - $ref: "#/components/parameters/Service"
        - $ref: "#/components/parameters/Host"
        - $ref: "#/components/parameters/Unit"
        - $ref: "#/components/parameters/Level"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Stream"
        - $ref: "#/components/parameters/TraceID"
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/OperationID"
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/OrgID"
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/Component"
        - $ref: "#/components/parameters/ActorID"
        - $ref: "#/components/parameters/ActorType"
        - $ref: "#/components/parameters/Outcome"
        - $ref: "#/components/parameters/StatusCode"
        - $ref: "#/components/parameters/StatusClass"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Order"
      responses:
        "200":
          description: Matching log events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: deployment-provisioned-token
      description: >
        Required when `IngestConfig.Token` is configured. The token value must
        be sent as `Authorization: Bearer <token>`.
  parameters:
    Since:
      name: since
      in: query
      description: Inclusive lower timestamp bound.
      schema:
        type: string
        format: date-time
    Until:
      name: until
      in: query
      description: Inclusive upper timestamp bound.
      schema:
        type: string
        format: date-time
    Env:
      name: env
      in: query
      schema:
        type: string
      example: staging
    EventID:
      name: event_id
      in: query
      schema:
        type: string
      example: 6f8d9cf1b41e4f4aa1e76f0a7c6328e8b8dbd48c2d5724ecacda54f1e0f0d942
    Service:
      name: service
      in: query
      schema:
        type: string
      example: video-cloud-api
    Host:
      name: host
      in: query
      schema:
        type: string
      example: api-1
    Unit:
      name: unit
      in: query
      schema:
        type: string
      example: video-cloud-api.service
    Level:
      name: level
      in: query
      schema:
        $ref: "#/components/schemas/LogLevel"
    Source:
      name: source
      in: query
      schema:
        type: string
    Stream:
      name: stream
      in: query
      description: Logical stream, such as `operational` or `billing_usage`.
      schema:
        type: string
      example: journald
    TraceID:
      name: trace_id
      in: query
      schema:
        type: string
    RequestID:
      name: request_id
      in: query
      schema:
        type: string
    OperationID:
      name: operation_id
      in: query
      schema:
        type: string
    DeviceID:
      name: device_id
      in: query
      schema:
        type: string
    OrgID:
      name: org_id
      in: query
      schema:
        type: string
    UserID:
      name: user_id
      in: query
      schema:
        type: string
    Component:
      name: component
      in: query
      schema:
        type: string
      example: device_runtime_log
    ActorID:
      name: actor_id
      in: query
      description: Actor identifier for admin audit and support drilldown.
      schema:
        type: string
    ActorType:
      name: actor_type
      in: query
      description: Actor class, such as `cloud_admin`, `service`, or `device`.
      schema:
        type: string
    Outcome:
      name: outcome
      in: query
      description: Stable operation result, such as `success` or `failure`.
      schema:
        type: string
    StatusCode:
      name: status_code
      in: query
      description: HTTP or operation status code represented as a string.
      schema:
        type: string
    StatusClass:
      name: status_class
      in: query
      description: Status class such as `2xx`, `4xx`, or `5xx`.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum number of events to return. Defaults to 100; maximum is 1000.
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Order:
      name: order
      in: query
      description: Timestamp ordering for returned events.
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
  responses:
    BadRequest:
      description: Invalid request body or query parameter.
      content:
        text/plain:
          schema:
            type: string
          example: invalid query parameter
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        text/plain:
          schema:
            type: string
          example: unauthorized
    MethodNotAllowed:
      description: HTTP method is not supported for this endpoint.
      content:
        text/plain:
          schema:
            type: string
          example: method not allowed
    PayloadTooLarge:
      description: Request body exceeds the configured ingest size limit.
      content:
        text/plain:
          schema:
            type: string
          example: request body too large
    InternalServerError:
      description: Backend query failed.
      content:
        text/plain:
          schema:
            type: string
          example: query failed
  schemas:
    IngestRequest:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/LogEvent"
    IngestResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/IngestResult"
    IngestResult:
      type: object
      required:
        - event_id
        - status
      properties:
        event_id:
          type: string
        status:
          type: string
          enum:
            - accepted
            - duplicate
            - rejected
        error:
          type: string
          description: Present when status is `rejected`.
    QueryResponse:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/LogEvent"
    LogEvent:
      type: object
      additionalProperties: false
      required:
        - event_id
        - ts
        - level
        - msg
        - service
        - env
        - version
        - host
        - unit
        - source
      properties:
        event_id:
          type: string
          description: Stable id generated from host_id, boot_id, unit, and cursor.
        ts:
          type: string
          format: date-time
        level:
          $ref: "#/components/schemas/LogLevel"
        msg:
          type: string
          description: Stable event message.
        service:
          type: string
        env:
          type: string
        version:
          type: string
        host:
          type: string
        unit:
          type: string
        source:
          type: string
          example: journald
        stream:
          type: string
          description: Logical stream used to isolate billing usage from operational logs.
        trace_id:
          type: string
        request_id:
          type: string
        operation_id:
          type: string
        device_id:
          type: string
        org_id:
          type: string
        user_id:
          type: string
        component:
          type: string
        error_category:
          type: string
        actor_id:
          type: string
          description: Actor identifier for admin audit and support drilldown.
        actor_type:
          type: string
          description: Actor class, such as `cloud_admin`, `service`, or `device`.
        outcome:
          type: string
          description: Stable operation result, such as `success` or `failure`.
        status_code:
          type: string
          description: HTTP or operation status code represented as a string.
        status_class:
          type: string
          description: Status class such as `2xx`, `4xx`, or `5xx`.
        fields:
          type: object
          additionalProperties: true
          description: >
            Extra service fields preserved after redaction. Sensitive field
            values are stored as `[REDACTED]`.
    LogLevel:
      type: string
      enum:
        - debug
        - info
        - warn
        - error
        - dpanic
        - panic
        - fatal